move utils to separate crate
