+++ crates/marine-macro-impl/Cargo.toml
@@ -23,3 +23,4 @@ uuid = { version = "0.8.2", features = ["v4"] }
 
 [dev-dependencies]
 pretty_assertions = "0.7.1"
+marine-macro-testing-utils = {path = "../macro-testing-utils"}
